moved verifiable-resentation-lib into identity-hub-core
