+++ core/identity-hub-core/src/main/java/org/eclipse/edc/identityhub/core/services/verifiablepresentation/generators/PresentationGeneratorConstants.java
@@ -1,0 +1,27 @@
+/*
+ *  Copyright (c) 2025 Cofinity-X
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Cofinity-X - initial API and implementation
+ */
+
+package org.eclipse.edc.identityhub.core.services.verifiablepresentation.generators;
+
+/**
+ * Contains common constants for {@link LdpPresentationGenerator} and {@link JwtPresentationGenerator}.
+ */
+public interface PresentationGeneratorConstants {
+
+    String CONTROLLER_ADDITIONAL_DATA = "controller";
+
+    String VP_TYPE_PROPERTY = "type";
+
+    String VERIFIABLE_CREDENTIAL_PROPERTY = "verifiableCredential";
+
+}
