+++ core/identity-hub-core/src/test/resources/credentials.v1.json
