+++ core/identity-hub-core/src/test/resources/dcp.v08.json
