+++ core/identity-hub-core/src/test/resources/did.json
