+++ core/identity-hub-core/src/test/resources/examples.v1.json
