+++ core/identity-hub-core/src/test/resources/jws2020.json
