+++ core/identity-hub-core/src/test/resources/presentation-exchange.v1.json
